Refactor ReplaceLicenseByClearlyDefined to use "Ref" instead of "Purl"
Replaced all instances of "Purl" with "Ref" in configurations, logic, and tests for consistency and clarity. Adjusted related method names, validation, and logging to align with the new "Ref" terminology. Ensured all test cases were updated to reflect the changes without altering existing functionality.
