+++ src/CdxEnrich.Tests/Actions/ReplaceLicenseByClearlyDefined/testcases/configs/invalid-git.yaml
@@ -1,2 +1,2 @@
 ﻿ReplaceLicenseByClearlyDefined:
-  - Purl: "pkg:git/github.com/expressjs/express@4.18.2"
+  - Ref: "pkg:git/github.com/expressjs/express@4.18.2"
