+++ src/CdxEnrich.Tests/Actions/ReplaceLicenseByClearlyDefined/testcases/configs/valid-all-supported.yaml
@@ -1,8 +1,8 @@
 ﻿ReplaceLicenseByClearlyDefined:
-  - Purl: "pkg:nuget/System.Collections@4.0.11"
-  - Purl: "pkg:npm/lodash@4.17.21"
-  - Purl: "pkg:maven/org.apache.commons/commons-lang3@3.12.0"
-  - Purl: "pkg:pypi/requests@2.28.1"
-  - Purl: "pkg:gem/rails@7.0.4"
-  - Purl: "pkg:pod/Alamofire@5.6.2"
-  - Purl: "pkg:crate/serde@1.0.152"
+  - Ref: "pkg:nuget/System.Collections@4.0.11"
+  - Ref: "pkg:npm/lodash@4.17.21"
+  - Ref: "pkg:maven/org.apache.commons/commons-lang3@3.12.0"
+  - Ref: "pkg:pypi/requests@2.28.1"
+  - Ref: "pkg:gem/rails@7.0.4"
+  - Ref: "pkg:pod/Alamofire@5.6.2"
+  - Ref: "pkg:crate/serde@1.0.152"
